finish role portion and add teacher manager

+++ src/main/java/com/njmsita/exam/authentic/service/ebo/RoleEbo.java
@@ -1,9 +1,14 @@
 package com.njmsita.exam.authentic.service.ebo;
 
 import com.njmsita.exam.authentic.dao.dao.RoleDao;
+import com.njmsita.exam.authentic.dao.dao.StudentDao;
+import com.njmsita.exam.authentic.dao.dao.TeacherDao;
+import com.njmsita.exam.authentic.model.StudentVo;
+import com.njmsita.exam.authentic.model.TeacherVo;
 import com.njmsita.exam.authentic.model.TroleVo;
 import com.njmsita.exam.authentic.service.ebi.RoleEbi;
 import com.njmsita.exam.base.BaseQueryVO;
+import com.njmsita.exam.manager.model.SchoolVo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,6 +25,10 @@ public class RoleEbo implements RoleEbi
 {
     @Autowired
     private RoleDao roleDao;
+    @Autowired
+    private TeacherDao teacherDao;
+    @Autowired
+    private StudentDao studentDao;
 
     public void save(TroleVo troleVo)
     {
@@ -54,6 +63,25 @@ public class RoleEbo implements RoleEbi
 
     public void delete(TroleVo roleVo)
     {
-        roleDao.delete(roleVo);
+        List<TeacherVo> teachers=teacherDao.getAllByRoleId(roleVo.getId());
+        List<TeacherVo> students=studentDao.getAllByRoleId(roleVo.getId());
+        if(0==teachers.size()&&students.size()==0){
+            roleDao.delete(roleVo);
+        }else{
+            //TODO 抛出异常
+            System.out.println("这个角色有关联的用户，不能删除");
+        }
+
+    }
+    //----------------------------------------------以上为基本方法------------------------------------------
+    //----------------------------------------------以上为基本方法------------------------------------------
+    //----------------------------------------------以上为基本方法------------------------------------------
+    //----------------------------------------------以上为基本方法------------------------------------------
+    //----------------------------------------------以上为基本方法------------------------------------------
+    //----------------------------------------------以上为基本方法------------------------------------------
+
+    public TroleVo getByName(String name)
+    {
+        return roleDao.getByName(name);
     }
 }
